services/routing/requirements.txt
django==5.1.4
djangorestframework==3.15.2
python-dotenv==1.0.1
numpy==2.2.0
qiskit==1.3.1
qiskit-algorithms==0.3.1
scipy==1.14.1